+++ mocks/aws-sdk/clients/dynamodb.js
@@ -1,10 +1,3 @@
-const S3 = jest.fn();
-S3.prototype = {
-  listObjects: jest.fn((params, callback) => {
-    callback(null, { objects: [] });
-  }),
-};
-
 const DynamoDB = jest.fn();
 DynamoDB.prototype = {
   scan: jest.fn((params, callback) => {
@@ -18,8 +11,4 @@ DynamoDB.prototype = {
   }),
 };
 
-export default {
-  CognitoIdentityCredentials: jest.fn(authInfo => authInfo),
-  DynamoDB,
-  S3,
-};
+export default DynamoDB;
