+++ mocks/aws-sdk/clients/s3.js
@@ -1,0 +1,8 @@
+const S3 = jest.fn();
+S3.prototype = {
+  listObjects: jest.fn((params, callback) => {
+    callback(null, { objects: [] });
+  }),
+};
+
+export default S3;
