+++ mocks/aws-sdk/global.js
@@ -1,0 +1,3 @@
+export default {
+  CognitoIdentityCredentials: jest.fn(authInfo => authInfo),
+};
