+++ src/services/s3.js
@@ -1,11 +1,12 @@
-import AWS from 'aws-sdk-umd';
+import AWS from 'aws-sdk/global';
+import AWSS3 from 'aws-sdk/clients/s3';
 import {
   BUCKET,
   REGION,
 } from '../constants/aws';
 
 export function getBucketContents(authInfo) {  // eslint-disable-line import/prefer-default-export
-  const S3 = new AWS.S3({
+  const S3 = new AWSS3({
     region: REGION,
     credentials: new AWS.CognitoIdentityCredentials(authInfo.params),
   });
